Enable file URIs for non-targz openshift files

+++ pkg/minikube/sshutil/sshutil.go
@@ -62,7 +62,7 @@ func NewSSHClient(d drivers.Driver) (*ssh.Client, error) {
 }
 
 // Transfer uses an SSH session to copy a file to the remote machine.
-func Transfer(reader io.Reader, readerLen int, remotedir, filename string, perm string, c *ssh.Client) error {
+func Transfer(reader io.Reader, readerLen int64, remotedir, filename string, perm string, c *ssh.Client) error {
 	// Delete the old file first. This makes sure permissions get reset.
 	deleteCmd := fmt.Sprintf("sudo rm -f %s", filepath.Join(remotedir, filename))
 	mkdirCmd := fmt.Sprintf("sudo mkdir -p %s", remotedir)
